refactor(web-ui): move the connect button into the Networks tab row
Relocate the global connect/disconnect toggle out of the app toolbar
(and its mobile drawer copy) and into the Networks view, right-aligned
on the ED2K/Kad tab strip (see amule-org/amule#402). Behaviour is
unchanged: same both-networks toggle, coloured plug states, confirm
dialog, and admin-only gating.

- components.js: Tabs gains an optional right-aligned `extra` slot,
  rendered inside the tab strip so it shares the bottom border line.
- networks.js: owns ConnectButton now (reads status via useStore) and
  passes it as `extra` to the top tabs.
- app.js: remove both toolbar ConnectButton renders, the component
  itself, the now-unused status var and confirmDialog import; the Kad
  status-bar default now reads "Not connected" to match ED2K.
- app.css: drop the three dead toolbar/drawer .conn-btn rules; add
  .tabs-extra styling so the button reads as a compact button on the
  tab strip.
- i18n (en/es): uppercase app_ed2k label; relabel the per-network
  buttons to "Disconnect ED2K", "Connect from known clients" and
  "Disconnect Kad".
