+++ src/webapi/static/i18n/en.json
@@ -23,7 +23,7 @@
   "app_connecting": "Connecting…",
   "app_disconnect": "Disconnect",
   "app_disconnected": "Disconnected",
-  "app_ed2k": "ed2k",
+  "app_ed2k": "ED2K",
   "app_error": "Error",
   "app_files": "Files",
   "app_firewalled": "Firewalled",
@@ -97,7 +97,7 @@
   "networks_server_confirm_remove": "Remove server “{name}”?",
   "networks_server_connect": "Connect",
   "networks_server_description": "Description",
-  "networks_server_disconnect": "Disconnect",
+  "networks_server_disconnect": "Disconnect ED2K",
   "networks_server_disconnected": "Disconnected",
   "networks_server_empty": "Nothing here yet.",
   "networks_server_error": "Error",
@@ -118,9 +118,9 @@
   "networks_server_version": "Version",
   "networks_kad_bootstrap_from_node": "Bootstrap from node",
   "networks_kad_buddy": "Buddy",
-  "networks_kad_connect": "Connect",
+  "networks_kad_connect": "Connect from known clients",
   "networks_kad_contacts_nodes": "Contacts (nodes)",
-  "networks_kad_disconnect": "Disconnect",
+  "networks_kad_disconnect": "Disconnect Kad",
   "networks_kad_error": "Error",
   "networks_kad_files": "Files",
   "networks_kad_firewalled_tcp": "Firewalled (TCP)",
