+++ src/webapi/static/i18n/es.json
@@ -23,7 +23,7 @@
   "app_connecting": "Conectando…",
   "app_disconnect": "Desconectar",
   "app_disconnected": "Desconectado",
-  "app_ed2k": "ed2k",
+  "app_ed2k": "ED2K",
   "app_error": "Error",
   "app_files": "Archivos",
   "app_firewalled": "Tras cortafuegos",
@@ -97,7 +97,7 @@
   "networks_server_confirm_remove": "¿Eliminar el servidor «{name}»?",
   "networks_server_connect": "Conectar",
   "networks_server_description": "Descripción",
-  "networks_server_disconnect": "Desconectar",
+  "networks_server_disconnect": "Desconectar ED2K",
   "networks_server_disconnected": "Desconectado",
   "networks_server_empty": "Aún no hay nada aquí.",
   "networks_server_error": "Error",
@@ -118,9 +118,9 @@
   "networks_server_version": "Versión",
   "networks_kad_bootstrap_from_node": "Bootstrap desde nodo",
   "networks_kad_buddy": "Compañero",
-  "networks_kad_connect": "Conectar",
+  "networks_kad_connect": "Conectar desde clientes conocidos",
   "networks_kad_contacts_nodes": "Contactos (nodos)",
-  "networks_kad_disconnect": "Desconectar",
+  "networks_kad_disconnect": "Desconectar Kad",
   "networks_kad_error": "Error",
   "networks_kad_files": "Archivos",
   "networks_kad_firewalled_tcp": "Tras cortafuegos (TCP)",
