+++ src/webapi/static/js/views/networks.js
@@ -2,9 +2,11 @@
 // top notebook with ED2K (server list) and Kad tabs, and a bottom notebook
 // with the aMule log and Server info. Composed from the existing panels.
 
-import { html, useState } from "../dom.js";
-import { Tabs } from "../components.js";
-import { t } from "../i18n.js";
+import { html, useState, useStore } from "../dom.js";
+import { api } from "../api.js";
+import { Tabs, toast, confirmDialog } from "../components.js";
+import { t, terr } from "../i18n.js";
+import { Icon } from "../icons.js";
 import { ServersPanel } from "./servers.js";
 import { KadPanel, KadInfoPanel } from "./kad.js";
 import { Ed2kInfoPanel } from "./ed2k.js";
@@ -28,7 +30,8 @@ export default function Networks({ isGuest }) {
   return html`
     <div class="net-split">
       <section class="net-pane">
-        <${Tabs} tabs=${topTabs} active=${top} onSelect=${setTop} />
+        <${Tabs} tabs=${topTabs} active=${top} onSelect=${setTop}
+                 extra=${html`<${ConnectButton} />`} />
         <div class="net-pane-body">
           ${top === "ed2k"
             ? html`<${ServersPanel} isGuest=${isGuest} />`
@@ -49,4 +52,43 @@ export default function Networks({ isGuest }) {
         </div>
       </section>
     </div>`;
+}
+
+// aMule-style connection button: a coloured plug (red = disconnected, amber =
+// connecting, green = connected) that toggles both networks. Lives on the
+// right of the ED2K/Kad tab row.
+function ConnectButton() {
+  const status = useStore("status");
+  const ed2k = (status && status.ed2k) || {};
+  const kad = (status && status.kad) || {};
+  const ed2kConn = ed2k.state === "connected";
+  // "connecting" wins over "connected": switching ed2k servers while Kad
+  // stays up (or vice versa) should still surface as a transition, not get
+  // masked by the other network already being connected.
+  const connecting = ed2k.state === "connecting" || kad.state === "connecting";
+  const connected = !connecting && (ed2kConn || kad.state === "connected");
+
+  // Colour and label follow the actual current state, not the action the
+  // click would perform.
+  const cls = connected ? "connected" : connecting ? "connecting" : "disconnected";
+  const label = t("app_" + cls);
+  const toggle = async () => {
+    try {
+      if (connected) {
+        if (!(await confirmDialog(t("app_confirm_disconnect_both"),
+              { okLabel: t("app_disconnect") }))) return;
+        await api.post("networks/disconnect", { network: "both" });
+        toast(t("app_toast_disconnecting"), "success");
+      } else {
+        await api.post("networks/connect", { network: "both" });
+        toast(t("app_toast_connecting"), "success");
+      }
+    } catch (e) { toast(terr(e) || t("app_error"), "error"); }
+  };
+
+  return html`
+    <button class=${"tool-btn conn-btn admin-only " + cls} title=${label} onClick=${toggle}>
+      <${Icon} name="connect" size=${20} />
+      <span class="tool-label">${label}</span>
+    </button>`;
 }
